Add CI to Pretty Screeplot

After CIs are determined by a combination of booted_eigenvalues & getCIs This function provided with the plot object and the data.frame with CIs adds CI around the observd eigenvalues

Usage

add_ci_2plot(
  plot,
  CIs,
  met = "normal",
  color = "darkseagreen3",
  outline = "black",
  transparency = 0.5,
  type = "band",
  lt = "dotted"
)

Arguments

plot

The plot object as obtained by pretty_scree()

CIs

Data.frame obtained from getCIs() and booted_eigenvalues()

met

the type of CI to be used default "normal" | other options: "basic", "stud", "perc", "bca"

color

If type of visual display of the CI = "band" the color of the area between Upper and Lower

outline

Color of the Outline for the Area

transparency

Alpha value, how transparent shall the filling color be

type

either "band" or "errorbars", defaults to "band"

lt

Linetype, 0 for no outline around the band, other options include: "dotted", "dashed", "solid"

Value

A new pretty_plot with the Parallel Analysis, including CIs around the observed Eigenvalues.

See also

Author

Bjoern Buedenbender