Creates a Descriptive Bivariate Table1 for Publication (Table1 + Flextable)

A convenience function, that provides and easy wrapper for the two main enginges of the function

table1 provides a nice API given a formula to create demographics tables. I basically just advanced the functionality of the p-value function, added the possibility to correct p-values with either Bonferroni or Sidark, and set some sensible defaults to achieve a nice look

flextable which gives all the power to format the table as you please (e.g., conditional formatting -> adding bold for p values below .05), adding italic headers or notes explaining what was done.

Really all credit should go to these two packages their developers, on full disclosure my function just provides an easy to use API or wrapper around their packages to get a beautiful publication ready bivariate comparison Table 1. New feature since version (0.2.3) comparisons with (Welch) ANOVA for more than 2 Groups

Usage

flex_table1(
  str_formula,
  data,
  correct = "bonf",
  num = NA,
  table_caption = NA,
  ref_correction = TRUE,
  ...
)

Arguments

str_formula

A string representing a formula, e.g., "~ Sepal.Length + Sepal.Width | Species" used to construct the table1.

data

The dataset containing the variables for the table1 call (all terms from the str_formula must be present)

correct

Character, currently available are "bonf" for Bonferroni correction or "sidark" for Sidark correction. Provide NA if you dont want any correction. Defaults to "bonf"

num

Integer number of comparisons. If NA will be determined automatically, by the number of terms in the formula

table_caption

Caption for the table, each element of the vector represents a new line. The first line will be bold face. All additional lines are in italic.

ref_correction

Boolean, default = TRUE, if TRUE corrected p-Values will be referenced in the foot note.

...

(Optional), Additional arguments that can be passed to format_flextable (e.g., fontsize, font ...) or to serialNext

Value

A flextable object with APA ready correlation table. If a filepath is provided it also creates the respective file (e.g., a word .docx file)

See also

Author

Bjoern Buedenbender

Examples

if (FALSE) {
# Comparison of just two Groups
str_formula <- "~ Sepal.Length + Sepal.Width +test | Species"
data <- dplyr::filter(iris, Species %in% c("setosa", "versicolor"))
data$test <- factor(rep(c("Female", "Male"), 50))
table_caption <- c("Table 1", "A test on the Iris Data")
flex_table1(str_formula, data = data, table_caption = table_caption)

# Comparison of Multiple Groups (ANOVA)
str_formula <- "~ Sepal.Length + Sepal.Width + Gender_example | Species"
data <- dplyr::filter(iris, Species %in% c("setosa", "versicolor"))
data <- iris
data$Gender_example <- factor(rep(c("Female", "Male"), nrow(data)/2))
table_caption <- c("Table 1", "A test on the Iris Data")
flex_table1(str_formula, data = data, table_caption = table_caption)
}