Get ICD-10 F-Codes as Character Vector

Returns a character vector of length 11 for all ICD-10 F Categories (mental disorders)

Usage

get_ICD_10_cats(lang = "eng")

Arguments

lang

Language for the character vector, curr available in english and german (lang = "ger"), Default: "eng"

Value

Character Vector containing the 11 categroies for mental disorders (F Codes F01-F99)

Author

Bjoern Buedenbender

Examples

get_ICD_10_cats() # returns english ICD-10 Cats
#>  [1] "F01-F09 Mental disorders due to known physiological conditions"                                      
#>  [2] "F10-F19 Mental and behavioral disorders due to psychoactive substance use"                           
#>  [3] "F20-F29 Schizophrenia, schizotypal, delusional, and other non-mood psychotic disorders"              
#>  [4] "F30-F39 Mood [affective] disorders"                                                                  
#>  [5] "F40-F48 Anxiety, dissociative, stress-related, somatoform and other nonpsychotic mental disorders"   
#>  [6] "F50-F59 Behavioral syndromes associated with physiological disturbances and physical factors"        
#>  [7] "F60-F69 Disorders of adult personality and behavior"                                                 
#>  [8] "F70-F79 Intellectual disabilities"                                                                   
#>  [9] "F80-F89 Pervasive and specific developmental disorders"                                              
#> [10] "F90-F98 Behavioral and emotional disorders with onset usually occurring in childhood and adolescence"
#> [11] "F99-F99 Unspecified mental disorder"                                                                 
get_ICD_10_cats("ger") # returns the german ones
#>  [1] "F01-F09 Organische, einschließlich symptomatischer psychischer Störungen"          
#>  [2] "F10-F19 Psychische und Verhaltensstörungen durch psychotrope Substanzen"           
#>  [3] "F20-F29 Schizophrenie, schizotype und wahnhafte Störungen"                         
#>  [4] "F30-F39 Affektive Störungen"                                                       
#>  [5] "F40-F48 Neurotische, Belastungs- und somatoforme Störungen"                        
#>  [6] "F50-F59 Verhaltensauffälligkeiten mit körperlichen Störungen und Faktoren"         
#>  [7] "F60-F69 Persönlichkeits- und Verhaltensstörungen"                                  
#>  [8] "F70-F79 Intelligenzstörung"                                                        
#>  [9] "F80-F89 Entwicklungsstörungen"                                                     
#> [10] "F90-F98 Verhaltens- und emotionale Störungen mit Beginn in der Kindheit und Jugend"
#> [11] "F99-F99 Nicht näher bezeichnete psychische Störungen"